refactor(runtime): split worker session observations
